+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Fixed
 - Custom status for processing was not set as expected.
+- Direct debit secured: invoice was not sent automatically.
+- Prepayment: Payment information did not appear on invoice.
 
 ## v19.4.16